+++ src/FunFair.CodeAnalysis/ClassAnalysisDiagnosticsAnalyzer.cs
@@ -1,0 +1,60 @@
+﻿using System.Collections.Immutable;
+using System.Linq;
+using Microsoft.CodeAnalysis;
+using Microsoft.CodeAnalysis.CSharp;
+using Microsoft.CodeAnalysis.CSharp.Syntax;
+using Microsoft.CodeAnalysis.Diagnostics;
+
+namespace FunFair.CodeAnalysis
+{
+    /// <summary>
+    ///     Looks for issue with class declarations
+    /// </summary>
+    [DiagnosticAnalyzer(LanguageNames.CSharp)]
+    public sealed class ClassAnalysisDiagnosticsAnalyzer : DiagnosticAnalyzer
+    {
+        private const string CATEGORY = "Classes";
+
+        private static readonly DiagnosticDescriptor Rule = RuleHelpers.CreateRule(Rules.RuleClassesShouldBeStaticSealedOrAbstract,
+                                                                                   CATEGORY,
+                                                                                   title: "Classes should be static, sealed or abstract",
+                                                                                   message: "Classes should be static, sealed or abstract");
+
+        /// <inheritdoc />
+        public override ImmutableArray<DiagnosticDescriptor> SupportedDiagnostics => new[] {Rule}.ToImmutableArray();
+
+        /// <inheritdoc />
+        public override void Initialize(AnalysisContext context)
+        {
+            context.ConfigureGeneratedCodeAnalysis(GeneratedCodeAnalysisFlags.Analyze | GeneratedCodeAnalysisFlags.None);
+            context.EnableConcurrentExecution();
+
+            context.RegisterCompilationStartAction(PerformCheck);
+        }
+
+        private static void PerformCheck(CompilationStartAnalysisContext compilationStartContext)
+        {
+            compilationStartContext.RegisterSyntaxNodeAction(MustBeReadOnly, SyntaxKind.ClassDeclaration);
+        }
+
+        private static void MustBeReadOnly(SyntaxNodeAnalysisContext syntaxNodeAnalysisContext)
+        {
+            if (!(syntaxNodeAnalysisContext.Node is ClassDeclarationSyntax classDeclarationSyntax))
+            {
+                return;
+            }
+
+            if (!classDeclarationSyntax.Modifiers.Any(IsWhiteListedClassModifier))
+            {
+                syntaxNodeAnalysisContext.ReportDiagnostic(Diagnostic.Create(Rule, classDeclarationSyntax.GetLocation()));
+            }
+        }
+
+        private static bool IsWhiteListedClassModifier(SyntaxToken syntaxToken)
+        {
+            SyntaxKind kind = syntaxToken.Kind();
+
+            return kind == SyntaxKind.StaticKeyword || kind == SyntaxKind.AbstractKeyword || kind == SyntaxKind.SealedKeyword;
+        }
+    }
+}